renamed "middleware" submodule > "editing"
renamed "edge" submodule > "middleware"
moved editing-data-service to "editing" submodule
EditingDataCache methods (and EditingDataDiskCache implementation) are now async (so that we can perform disk cleanup after Promise resolve)
implemented BasicEditingDataService, more adjustments to ServerlessEditingDataService
