+++ packages/create-sitecore-jss/src/templates/nextjs-sxa/src/lib/middleware/plugins/redirects.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { RedirectsMiddleware } from '@sitecore-jss/sitecore-jss-nextjs/edge';
+import { RedirectsMiddleware } from '@sitecore-jss/sitecore-jss-nextjs/middleware';
 import config from 'temp/config';
 import { MiddlewarePlugin } from '..';
