+++ packages/create-sitecore-jss/src/templates/nextjs/src/lib/page-props-factory/plugins/preview-mode.ts
@@ -1,4 +1,4 @@
-import { editingDataService } from '@sitecore-jss/sitecore-jss-nextjs';
+import { editingDataService } from '@sitecore-jss/sitecore-jss-nextjs/editing';
 import { SitecorePageProps } from 'lib/page-props';
 import { GetServerSidePropsContext, GetStaticPropsContext } from 'next';
 import { Plugin } from '..';
