+++ packages/create-sitecore-jss/src/templates/nextjs/src/pages/api/editing/data/[key].ts
@@ -1,20 +1,21 @@
-import { EditingDataMiddleware } from '@sitecore-jss/sitecore-jss-nextjs/middleware';
+import { EditingDataMiddleware } from '@sitecore-jss/sitecore-jss-nextjs/editing';
 
 /**
  * This Next.js API route is used to handle Sitecore editor data storage and retrieval by key
- * (between the Sitecore editor POST and Next.js page render requests) via the `EditingDataService`.
+ * on serverless deployment architectures (e.g. Vercel) via the `ServerlessEditingDataService`.
  *
  * The `EditingDataMiddleware` expects this dynamic route name to be '[key]' by default, but can
  * be configured to use something else with the `dynamicRouteKey` option.
  */
 
-// Bump body size limit (1mb by default) for Sitecore editor payload
-// See https://nextjs.org/docs/api-routes/api-middlewares#custom-config
+// Bump body size limit (1mb by default) and disable response limit for Sitecore editor payloads
+// See https://nextjs.org/docs/api-routes/request-helpers#custom-config
 export const config = {
   api: {
     bodyParser: {
       sizeLimit: '2mb',
     },
+    responseLimit: false,
   },
 };
