--- packages/sitecore-jss-nextjs/edge.d.ts
@@ -1,1 +1,0 @@
-export * from './types/edge/index';
