--- packages/sitecore-jss-nextjs/edge.js
@@ -1,1 +1,0 @@
-module.exports = require('./dist/cjs/edge/index');
