+++ packages/sitecore-jss-nextjs/editing.d.ts
@@ -1,0 +1,1 @@
+export * from './types/editing/index';
