+++ packages/sitecore-jss-nextjs/editing.js
@@ -1,0 +1,1 @@
+module.exports = require('./dist/cjs/editing/index');
