--- packages/sitecore-jss-nextjs/src/edge/index.ts
@@ -1,2 +1,0 @@
-export { RedirectsMiddleware, RedirectsMiddlewareConfig } from './redirects-middleware';
-export { PersonalizeMiddleware, PersonalizeMiddlewareConfig } from '../edge/personalize-middleware';
