+++ packages/sitecore-jss-nextjs/src/editing/editing-data-cache.ts
@@ -2,14 +2,14 @@
 /// <reference types="../../global" />
 import Cache, { CacheInstance } from 'sync-disk-cache';
 import os from 'os';
-import { EditingData } from '../sharedTypes/editing-data';
+import { EditingData } from './editing-data';
 
 /**
  * Defines an editing data cache implementation
  */
 export interface EditingDataCache {
-  set(key: string, editingData: EditingData): void;
-  get(key: string): EditingData | undefined;
+  set(key: string, editingData: EditingData): Promise<void>;
+  get(key: string): Promise<EditingData | undefined>;
 }
 
 /**
@@ -27,23 +27,28 @@ export class EditingDataDiskCache implements EditingDataCache {
     this.cache = new Cache('editing-data', { compression: 'gzip', location: tmpDir });
   }
 
-  set(key: string, editingData: EditingData): void {
-    const filePath = this.cache.set(key, JSON.stringify(editingData));
-    if (!filePath || filePath.length === 0) {
-      throw new Error(`Editing data cache not set for key ${key} at ${this.cache.root}`);
-    }
+  set(key: string, editingData: EditingData): Promise<void> {
+    return new Promise((resolve, reject) => {
+      const filePath = this.cache.set(key, JSON.stringify(editingData));
+      if (!filePath || filePath.length === 0) {
+        reject(`Editing data cache not set for key ${key} at ${this.cache.root}`);
+      }
+      resolve();
+    });
   }
 
-  get(key: string): EditingData | undefined {
-    const entry = this.cache.get(key);
-    if (!entry.isCached) {
-      console.warn(`Editing data cache miss for key ${key} at ${this.cache.root}`);
-      return undefined;
-    }
-    const data = JSON.parse(entry.value);
-    // Remove immediately to preserve disk-space
-    this.cache.remove(key);
-    return data as EditingData;
+  get(key: string): Promise<EditingData | undefined> {
+    return new Promise((resolve) => {
+      const entry = this.cache.get(key);
+      if (!entry.isCached) {
+        console.warn(`Editing data cache miss for key ${key} at ${this.cache.root}`);
+        resolve(undefined);
+      }
+      const data = JSON.parse(entry.value);
+      resolve(data as EditingData);
+      // Remove immediately to preserve disk-space
+      this.cache.remove(key);
+    });
   }
 }
