+++ packages/sitecore-jss-nextjs/src/editing/index.ts
@@ -1,0 +1,15 @@
+export { EditingData } from './editing-data';
+export { EditingDataCache, EditingDataDiskCache } from './editing-data-cache';
+export { EditingDataMiddleware, EditingDataMiddlewareConfig } from './editing-data-middleware';
+export {
+  EditingRenderMiddleware,
+  EditingRenderMiddlewareConfig,
+} from './editing-render-middleware';
+export {
+  EditingPreviewData,
+  EditingDataService,
+  BasicEditingDataService,
+  ServerlessEditingDataService,
+  ServerlessEditingDataServiceConfig,
+  editingDataService,
+} from './editing-data-service';
