+++ packages/sitecore-jss-nextjs/src/middleware/index.ts
@@ -1,6 +1,2 @@
-export { EditingDataCache } from './editing-data-cache';
-export { EditingDataMiddleware, EditingDataMiddlewareConfig } from './editing-data-middleware';
-export {
-  EditingRenderMiddleware,
-  EditingRenderMiddlewareConfig,
-} from './editing-render-middleware';
+export { RedirectsMiddleware, RedirectsMiddlewareConfig } from './redirects-middleware';
+export { PersonalizeMiddleware, PersonalizeMiddlewareConfig } from './personalize-middleware';
