+++ packages/sitecore-jss-nextjs/src/middleware/redirects-middleware.ts
